+++ src/rrt.c
@@ -177,7 +177,7 @@ void ferRRTDumpSVT(fer_rrt_t *rrt, FILE *out, const char *name)
         n  = fer_container_of(nn, fer_rrt_node_t, node);
         id2 = n->_id;
 
-        fprintf(out, "%d %d\n", id1, id2);
+        fprintf(out, "%d %d\n", (int)id1, (int)id2);
     }
 
     fprintf(out, "--------\n");
